Remove approvals from main
