+++ package.json
@@ -70,7 +70,7 @@
   "resolutions": {
     "**/json-schema": "^0.4.0",
     "@sentry/react-native/**/xmldom": "^0.6.0",
-    "@walletconnect/**/ws": "^7.4.6",
+
     "react-native-svg/**/nth-check": "^2.0.1",
     "**/set-value": "^4.0.1",
     "**/pac-resolver": "^5.0.0",
@@ -114,8 +114,8 @@
     "@sentry/integrations": "6.3.1",
     "@sentry/react-native": "2.4.2",
     "@tradle/react-native-http": "2.0.1",
-    "@walletconnect/client": "^1.5.1",
-    "@walletconnect/utils": "^1.5.1",
+    "@walletconnect/client": "^1.7.1",
+    "@walletconnect/utils": "^1.7.1",
     "asyncstorage-down": "4.2.0",
     "axios": "^0.21.2",
     "babel-plugin-transform-inline-environment-variables": "0.4.3",
